% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_G-SAM.R
\docType{data}
\name{G.SAM.corrected.DE.genes.VST}
\alias{G.SAM.corrected.DE.genes.VST}
\title{Real-world expression data from G-SAM project}
\format{
An object of class \code{data.frame} with 422 rows and 287 columns.
}
\usage{
G.SAM.corrected.DE.genes.VST
}
\description{
Real-world expression data from G-SAM project
}
\keyword{datasets}
