% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_GSAM.R
\docType{data}
\name{G.SAM.corrected.DE.labels}
\alias{G.SAM.corrected.DE.labels}
\title{Real-world expression data labels}
\format{
An object of class \code{data.frame} with 422 rows and 19 columns.
}
\usage{
G.SAM.corrected.DE.labels
}
\description{
Real-world expression data labels
}
\keyword{datasets}
