% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursiveCorPlot.R
\name{recursiveCorPlot-package}
\alias{recursiveCorPlot-package}
\title{recursiveCorPlot: natural clustering of RNA-seq data}
\description{
recursiveCorPlot provides two main functions:
recursiveCorPlot() and geom_circle()
}
\details{
See the vignette for more details, a step by step explanation of these functions.
}
