% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursiveCorPlot.R
\name{recursiveCorPlot}
\alias{recursiveCorPlot}
\title{recursiveCorPlot - corplot}
\usage{
recursiveCorPlot(
  normalised_correlated_data,
  labels,
  font_scale,
  legend_scale,
  method = "ward.D2",
  return_h_object = FALSE
)
}
\arguments{
\item{labels}{Matching labels (T, F, NA) for the genes (rownames set to same genes)}

\item{font_scale}{size of font}

\item{legend_scale}{size of legend blocks}

\item{method}{hclust method (see hclust for help)}

\item{return_h_object}{when T, return the h-clust object}

\item{normalised_correlation_data}{VST transformed or TMP read count table (rownames set to genes)}
}
\value{
nothing, or h-clust object if return_h_object was set to TRUE
}
\description{
This is an example function named 'recursiveCorPlot'
which makes a clustered correlation plot using recursive correlation as
distance metric.
Some useful keyboard shortcuts for package authoring:
}
