% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_recursiveCorPlot.R,
%   R/recursiveCorPlot.R
\docType{package}
\name{recursiveCorPlot}
\alias{recursiveCorPlot}
\title{recursiveCorPlot - corplot}
\usage{
recursiveCorPlot(
  normalised_correlated_data,
  labels,
  font_scale,
  legend_scale,
  method = "ward.D2",
  return_h_object = FALSE,
  caption = NULL
)
}
\arguments{
\item{normalised_correlated_data}{VST transformed or TMP read count table (rownames set to genes)}

\item{labels}{Matching labels (T, F, NA) for the genes (rownames set to same genes)}

\item{font_scale}{size of font}

\item{legend_scale}{size of legend blocks}

\item{method}{hclust method (see hclust for help)}

\item{return_h_object}{when T, return the h-clust object}

\item{caption}{caption (string) to be included}
}
\value{
nothing, or h-clust object if return_h_object was set to TRUE
}
\description{
recursiveCorPlot provides two main functions:
recursiveCorPlot() and geom_circle()
}
\details{
See the vignette for more details, a step by step explanation of these functions.
}
